Continue division between regular and keyless daos
