+++ src/test/java/org/hrorm/KeylessTest.java
@@ -63,5 +63,5 @@ public class KeylessTest {
         }
     }
 
-    
+
 }
